requirements.txt
flask==2.0.1
transformers==4.30.2
torch==2.0.1
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.26.0
pytest==7.0.1
pytest-flask==1.2.0 